Implement a red border around the scroll coordinates
This is useful, but slightly misleading since it's based on the current vram address. It moves around during the render relative to where the PPU is currently reading from the nametable, which is kind of useful, but not at all what you might expect.

While it reads and uses fine_x from the PPU, it essentially discards it unless you're mid-render, as the PPU clears out its internal fine_x register by the end of the frame.