Adjust default window spawn positions to be a bit more conservative
This is mostly for my own debugging purposes. This configuration causes the default-sized game window and exactly one debug panel to always fit nicely on a 720p monitor, which is my personal setup for testing. I think this is a reasonable default for now, but need to test on multiple OSes.