Add reset and tap commands
Reset is obvious, and resets the emulated NES.

Tap is for button inputs. While it's not feature complete or complex enough to be used like a movie, it can navigate menus in homebrew tests reasonably well, which is its purpose. Example:
tap start 5

The above will press the start key, emulate 5 frames, then release the start key.