Rip nearly all SDL references out of game / debug windows
This solves two problems. Firstly, it moves the debug UI panels closer
to being backend agnostic, so that they are more portable. Secondly,
it resolves an issue with canvas and texture generator lifetimes that
I couldn't figure out how to solve proberly before, allowing the
canvas texture to be created just once rather than every frame. This
represents a massive (60% CPU usage) performance gain on my system.

Main is a bit of a copy/pasta mess right now and due for an overhaul.