Add a piano roll debug window, to visualize music in realtime
Presently only works for the two pulse channels and the triangle
channel, but this is enough to view *most* "notes" in NES games.
The two missing channels (noise and DMC) are more difficult to
display in this format, but I'd like to figure something out for
them.

The piano roll displays the notes as they would be played on an
imaginary 76-key piano, with a range from A1 to C8, which roughly
matches what the NES is capable of playing, minus some of the
extremely high frequencies. Brightness corresponds to note volume.
Due both to accuracy limitations and occasionally intentional
de-tuning by the game, not all notes will be "centered" on their
given key, so you're able to see sharpness, flatness, and some
special effects like the output of the sweep register.