Draw fake DMC amplitude, tweak colors for consistency
This takes the difference between DMC samples (at 60 Hz, so not
every sample is considered) and draws the line width. This isn't
a real amplitude calculation, but it's close enough to suggest
one as a visualization. Most importantly, it accomplishes the goal
of easily distinguishing samples in the piano roll at a glance.