Create a TestWindow and loosely hook it up as a proof of concept
This works! It's missing any sort of event handling, but it does
process updates and drawing, so it's a decent start.

Side note: this is a whole bunch of duplicated code just to add one
new window. We very much need to reorganize the SDL build so that
these large repeated code blocks are moved into named functions.
Previously the hesitation revolved around a poor understanding of
lifetimes in Rust, but I don't think that's an issue anymore.