Wrap up Panels, Windows, and TextureCreators for iteration
Turns out the SDL2 library is *picky* when it comes to texture
creators and the resulting textures. The lifetimes are tricky
to wrangle, so here I've opted to keep the list of panels and
windows, texture creators, and textures all stuffed into
separate structs, with a synchronized index. This is the only
way I've found to keep Rust's borrow checker happy that doesn't
involve creating new textures every frame (very slow).