Display half the samples, for twice the apparent width
This is an aesthetic preference really, but I find this makes
especially melodic and pulse triangles much easier to read.

Additionally, add one extra guard to the last_y parameter, which
was causing out of bounds draws for a few games. (The final
buffer in particular is a little bit hand-wavey w/ respct to its
scaling, so occasional clipping isn't surprising.)