Hookup basic Standard Controller input translation
This code is doing a bit of work that I would rather not happen
here; in particular, an abstraction is needed for the I/O ports,
so that StandardController connections can be abstracted away.
Still, for the purposes of porting away from SDL manually handling
all of this, it's fine for now.