Implement gradients for Pulse Duty, matching Lua colors for now
The gradients here are basically being used as a flat list of
colors for the 2A03 case, but smaller numbers of colors are
supported, meaning a simple two-tone gradient is enough for most
timbres. The 2A03 is a special case because it essentially has
a *repeat* in the form of its 25% and 75% cycles, so it needs
the full set to represent this.