Add N163 support to the APU window and Piano Roll
Notable changes include: support for PatchIndex Timbre types, and
support for the APU window to dynamically resize *after* a game
has started playing audio, to handle cases where a game switches
N163 channel counts on the fly. The vast majority of software will
set just one channel count and stick to that, so a window resize
is generally okay here.