Update sdl2 to 0.34.5
This primarily fixes an issue with the audio queue, which on my systems was failing to completely empty under the old library version. This led the emulator to freeze shortly after loading any program, as it was never under the threshold to run the next frame.

More testing is needed; I don't completely trust this just because it runs on my system.