Insert a busy sleep on non-rendering frames
The crude sync to audio approach resulted in an unfortunate busy
loop, locking CPU usage to 100% on a single core. On non-rendering
frames, after we catch up audio and emulation, we insert a tiny
sleep; some testing revealed 1/10th of a millisecond to be
imperceptible, maintaining smooth updates for video, but dropping
CPU usage considerably; down to ~56% on my laptop.