Rework the NSF export script to pipe video directly to ffmpeg
This prevents us from needing to store the quite large uncompressed
frames on disk, which is far more convenient (not requiring users
to have crazy amounts of disk space) and also a performance boost
(not waiting on potentially very slow spinning disks to commit
the data). Now the performance is gated primarily by the rendering
speed of RusticNES's Piano Roll, and the video encoding speed of
ffmpeg operating on video frames in RAM.

Note that we intentionally do not apply this technique to audio,
since the disk usage and speed are much less of a concern. It turns
out that trying to feed two pipes into the ffmpeg encoder is rather
unstable, but a single pipe seems to work consistently.