Do NOT apply loudness normalization for now
loudnorm with the provided settings was causing the mix to vary
considerably, opposite of what we want for a music tool. It's possible
to apply linear normalization, which is what I wanted, but it requires
some more advanced setup and should really be guarded by an optional
flag. I'll do this later; for now I just want to remove the botched
variable normalization, as it's unfit for purpose here.