Have the video output include an alpha channel
The basic script converts to x264, immediately dropping this data,
but one of my custom scripts requires this and has a different
path. Eventually I'll update the base script to either output to
a more common format that supports an alpha channel, or allow the
user to select a codec.