Choose a default height that can actually hold all of the channels
As it turns out NSF files permit just *silly* amounts of channels,
and this can overwhelm the drawing canvas. I should probably debug
the order in which these resizes happen to fix the bug properly,
but frankly I'm likely to remove this panel instead, as it has
been thoroughly replaced by the piano roll view.