Update SDL2 to the latest version (it didn't fix the issue)
Turns out SDL2 on wayland is just weird in a lot of ways, and
I don't currently have the energy to investigate and figure
out why. Gonna ignore it for now and focus on the replacement
shell, but I might swing back around and dig in properly at
some point. All signs indicate that an upgrade to SDL3 might be
necessary to properly set flags for DPI scaling, among other
things.