Don't panic when the worker channel disconnects. It's fine.
We'll leave the panic on the main thread end because, during
a normal application shutdown, it should never be checked. But
if the worker disconnects during shutdown that's fairly normal.

Well, for now. We might want a graceful shutdown procedure that
has the threads clean themselves up, but for the moment that's
not necessary. Once we're done saving SRAM, there's no more state
to persist in the worker that we care about.