Hookup clicky-stuffs for video settings
This also sets the general framework and event flow for other
settings. Basically the egui frontend will operate on a cache of
values to display, while the application state has the true value
of everything and sends out events accordingly.

This is maybe a *bit* inefficient, as the cache will update for
each individual setting value during load, but that's a one-time
thing.