Begin work on a cleanroom port for the UI, isolate emu core functions
This is still a bit unclean, but it's a "builds and runs" starting
point. Now I want to isolate the emulator core even harder into a
module, so we aren't polluting the global scope and have a much
harder divided. The core has a lot of bonus threads and global state
at the moment.