Implement proper click to play activation for the audio stream
Also this detects when audio is stuck and attempts to recover, which
should help iOS users and other platforms where tabs can suspend while
not focused. Ideally we'd have background operation for the NSF player
but we can't really control how the platform works. We need to handle
every case as gracefully as we can.